Trying to fix errors.
Did a lot of small things: Renamed Binoperator to BinOperator and Unaryoperator to UnaryOperator. Parserwrapper scanner and parser now take a string path as input. Updated import for typechecker to match all abstractsyntax statements, expressions... Edited .bat file to only run on testtypechecker. for now. Edited test.java to work with boltparser.scanner and parser, not cocor.scanner and parser.